sql-migration-agents/requirements.txt
# Core dependencies
openai==1.3.5
sqlparse==0.4.4
pydantic==2.4.2

# Azure dependencies
azure-identity==1.13.0
azure-keyvault-secrets==4.7.0
azure-storage-blob==12.17.0
applicationinsights==0.11.11

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
loguru==0.7.0

# Testing
pytest==7.4.2
pytest-mock==3.11.1
pytest-asyncio==0.21.1

# Development
black==23.9.1
flake8==6.1.0 